cgdcbxd: populate stacked device priomap
APP data applies to the underlying device and any stacked devices
above it. This patch is required because skb's are typically orphaned
as soon as possible to improve socket memory management and reduce
cache misses. The easiest example of a stacked device that needs
this is the VLAN device.

This makes the assumption that by putting an application in a
cgdcbxd managed group you want it to managed per 802.1Qaz spec.
Meaning we make no attempt to track user changes in these directories
and cgdcbxd will overwrite any changes to real devs or stacked
devices.

Bonding case still needs to be handled it doesn't use iflink
and may be stacked on top of multiple devices.

We manage stacked devices by adding another socket to listen for
LINK events. When a link event is received we walk the list of
managed cgdcbxd_iface structs and virtual devices attached to
the iface struct for any ifindex values that match the new links
iflink.

I will handle the deletion case in a follow up patch. This is
broke for virtual devices and real devices and independent from
handling stacked interfaces. Also it appears that lldpad is not
removing app data in CEE mode in all cases.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>