cgdcbxd: assign default priority correctly in cgroups
When application data is deleted the netprio value should be set
to the default priority not zero. This also means when we get
a default priority app tlv we should configure the net_prio
values for any unset types.

This patch adds a dflt_priority variable per interface that
is tracked and set for any cgroup that does not have an
explicit application value. Further we scan each cgroup
on any event for the existence of any application data. If
no application data exists the group will be removed.

With this new tracking there is no need to keep the 'purge'
value which was being used to indicate directories should be
removed. Therefor we drop the variable from many of the
function routines.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>